add usePosition, fix PR issues
