+++ src/scenes/projects/projects.tsx
@@ -1,34 +1,13 @@
-import React, { useEffect, useRef, useState } from "react";
+import React from "react";
 import styles from "./projects.module.scss";
 import { useNavContext } from "../../data/navProvider";
+import { usePosition } from "../../hooks/usePosition";
 
 export function Projects() {
-  const [position, setPosition] = useState(0);
   const { updateProjectsPosition } = useNavContext();
-  const projectsRef = useRef<HTMLOptionElement | null>(null);
-
-  const handleResize = () => {
-    if (projectsRef.current) {
-      const newPosition = projectsRef.current.offsetTop;
-      setPosition(newPosition);
-    }
-  };
-  useEffect(() => {
-    window.addEventListener("resize", handleResize, { passive: true });
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  useEffect(() => handleResize(), []);
-  useEffect(() => updateProjectsPosition(position), [position]);
+  const { ref } = usePosition(updateProjectsPosition);
 
   return (
-    <section
-      id="projects"
-      ref={projectsRef}
-      className={styles.container}
-    ></section>
+    <section id="projects" ref={ref} className={styles.container}></section>
   );
 }
